feat: added resource filter to cluster diff
