+++ src/components/molecules/index.ts
@@ -8,3 +8,4 @@ export {default as ScrollIntoView} from './ScrollIntoView';
 export {default as TitleBar} from './TitleBar';
 export {default as SectionRenderer} from './SectionRenderer';
 export {default as NavigatorDiff} from './NavigatorDiff';
+export {default as ResourceFilter} from './ResourceFilter';
