+++ src/components/organisms/NavigatorPane/NavigatorPane.tsx
@@ -7,14 +7,13 @@ import HelmChartSectionBlueprint, {HelmChartScopeType} from '@src/navsections/He
 import K8sResourceSectionBlueprint, {K8sResourceScopeType} from '@src/navsections/K8sResourceSectionBlueprint';
 import {Popover} from 'antd';
 import {PlusOutlined, FilterOutlined} from '@ant-design/icons';
-import ResourceFilter from '@components/molecules/ResourceFilter';
 import {useAppDispatch, useAppSelector} from '@redux/hooks';
 import {NAVIGATOR_HEIGHT_OFFSET, ROOT_FILE_ENTRY} from '@constants/constants';
 import {useSelector} from 'react-redux';
 import {isInClusterModeSelector, isInPreviewModeSelector} from '@redux/selectors';
 import {openNewResourceWizard} from '@redux/reducers/ui';
 import {MonoPaneTitle} from '@components/atoms';
-import {SectionRenderer} from '@components/molecules';
+import {SectionRenderer, ResourceFilter} from '@components/molecules';
 import WarningsAndErrorsDisplay from './WarningsAndErrorsDisplay';
 import * as S from './NavigatorPane.styled';
